+++ pmxbot/slack.py
@@ -4,7 +4,7 @@ import importlib
 import pmxbot
 
 
-class Bot:
+class Bot(pmxbot.core.Bot):
 	def __init__(self, server, port, nickname, channels, password=None):
 		token = pmxbot.config['slack token']
 		sc = importlib.import_module('slackclient')
@@ -20,7 +20,11 @@ class Bot:
 			time.sleep(0.1)
 
 	def handle_message(self, msg):
-		"stubbed"
+		if msg.get('type') != 'message':
+			return
+		channel = self.client.server.channels.find(msg['channel']).name
+		nick = self.client.server.users.find(msg['user']).name
+		self.handle_action(channel, nick, msg['text'])
 
 	def handle_scheduled_tasks(self):
 		"stubbed"